requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
braceexpand==0.1.7
certifi==2025.8.3
charset-normalizer==3.4.2
clip @ git+https://github.com/ultralytics/CLIP.git@a6238961f9fab312d99c7709f8c7485b3f9a9b50
clip-benchmark==1.6.2
contourpy==1.3.3
cycler==0.12.1
datasets==4.0.0
dill==0.3.8
diskcache==5.6.3
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.3.0
ftfy==6.3.1
hf-xet==1.1.5
huggingface-hub==0.34.3
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
# llama_cpp_python==0.3.14
MarkupSafe==3.0.2
matplotlib==3.10.5
mobileclip @ git+https://github.com/ultralytics/mobileclip.git@21653a924a91094f3c2bd5e8e4203800a954d846
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
networkx==3.5
numpy==2.2.6
open_clip_torch==3.0.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow==11.3.0
propcache==0.3.2
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pycocoevalcap==1.2
pycocotools==2.0.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.19
tokenizers==0.21.4
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.54.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.173
ultralytics-thop==2.0.14
urllib3==2.5.0
wcwidth==0.2.13
webdataset==1.0.2
xxhash==3.5.0
yarl==1.20.1
